feat(cards): implementar validações de services, entities e controllers

+++ src/main/java/com/example/bankopsapi/service/CardService.java
@@ -1,8 +1,7 @@
 package com.example.bankopsapi.service;
 
 import com.example.bankopsapi.dto.request.CardRequestDTO;
-import com.example.bankopsapi.exception.CardNotFoundException;
-import com.example.bankopsapi.exception.IssuerNotFoundException;
+import com.example.bankopsapi.exception.*;
 import com.example.bankopsapi.model.Card;
 import com.example.bankopsapi.model.Issuer;
 import com.example.bankopsapi.repository.CardRepository;
@@ -20,6 +19,10 @@ public class CardService {
     private final IssuerRepository issuerRepository;
 
     public Card createCard(CardRequestDTO request) {
+        if (cardRepository.existsById(request.issuerId())){
+            throw new ExistsCardIssuerException();
+        }
+
         Issuer issuer = issuerRepository.findById(request.issuerId())
                 .orElseThrow(IssuerNotFoundException::new);
 
@@ -39,15 +42,33 @@ public class CardService {
     }
 
     public List<Card> listAllCards() {
-        return cardRepository.findAll();
+        List<Card> cards = cardRepository.findAll();
+
+        if (cards.isEmpty()) {
+            throw new NoCardFoundException();
+        }
+
+        return cards;
     }
 
     public Card findById(Long id) {
+        if (id == null || id <= 0){
+            throw new InvalidCardIdException();
+        }
+
         return cardRepository.findById(id)
                 .orElseThrow(CardNotFoundException::new);
     }
 
     public void deleteCard(Long id) {
+        if (id == null || id <= 0){
+            throw new InvalidCardIdException();
+        }
+
+        if (!cardRepository.existsById(id)) {
+            throw new InvalidCardIdException();
+        }
+
         cardRepository.deleteById(id);
     }
 }
